Made a connection to the database. Pulled up one servlet.

+++ src/main/java/repository/DiscountCardRepository.java
@@ -4,6 +4,6 @@ import model.DiscountCard;
 
 import java.util.Optional;
 
-public interface DiscountCardRepository extends Repository<DiscountCard, Long> {
+public interface DiscountCardRepository extends Repository<DiscountCard, Integer> {
     Optional<DiscountCard> getByNameDiscountCard(String numberCard);
 }
